Updated with latest versions

+++ scenes/objects/background.py
@@ -3,7 +3,7 @@ Sprite that holds background
 """
 
 import pygame
-from vector import Vec2d
+from .vector import Vec2d
 vec = Vec2d
 
 class BackgroundSprite(pygame.sprite.Sprite):
@@ -20,8 +20,8 @@ class BackgroundSprite(pygame.sprite.Sprite):
 
     def updateImage(self):
         self.image.fill((0,0,0))
-        for x in range(self.screenSize[0]/self.imageSize[0] + 1):
-            for y in range(-2, self.screenSize[1]/self.imageSize[1] + 3):
+        for x in range(int(self.screenSize[0]/self.imageSize[0] + 1.0)):
+            for y in range(-2, int(self.screenSize[1]/self.imageSize[1] + 3)):
                 self.image.blit(self.backgroundImage,
                     (self.pos.x + x*self.imageSize[0],self.pos.y + y*self.imageSize[1]))
